+++ src/array/codec/bytes_to_bytes.rs
@@ -2,6 +2,8 @@
 
 #[cfg(feature = "blosc")]
 pub mod blosc;
+#[cfg(feature = "bz2")]
+pub mod bz2;
 #[cfg(feature = "crc32c")]
 pub mod crc32c;
 #[cfg(feature = "gzip")]